+++ src/Console/.php-cs-fixer.dist.php
@@ -13,7 +13,6 @@ return (new PhpCsFixer\Config())
         '@PSR12' => true,
         'array_syntax' => ['syntax' => 'short'],
         'blank_line_after_opening_tag' => true,
-        'braces' => ['allow_single_line_closure' => true],
         'compact_nullable_typehint' => true,
         'concat_space' => ['spacing' => 'one'],
         'declare_equal_normalize' => ['space' => 'none'],
@@ -81,7 +80,7 @@ return (new PhpCsFixer\Config())
         'return_type_declaration' => ['space_before' => 'none'],
         'single_quote' => true,
         'single_trait_insert_per_statement' => true,
-        'standardize_not_equals' => true,
+        'standardize_not_equals' => true
     ])
     ->setRiskyAllowed(true)
     ->setFinder(